Add files via upload
mcroUnfold is used as main for now and it works with macro_table and util files. edited bot mcroUnfold and asmblrMain to also deal with filename extensions. can see mcroUnfold now as: deposit a file pointer and its name as argument, receive back a file.am with unfolded macros if passed the tests: all lines are equal or less 80 chars, all macros are declared once and not more, macro names are not using reserved words of the language.